face-auth/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pymongo==4.8.0
python-dotenv==1.0.1
cloudinary==1.41.0
cryptography==43.0.0
pillow==10.4.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
